feat: Implement cancel attempt workflow
